# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Circuvent Device API",
    "version": "1.0.0",
    "summary": "Read device state, send commands, and receive signed webhooks from Circuvent smart-home hardware.",
    "description": "The public API for the Circuvent control plane — the same one our own console and mobile apps use.\n\nAuthenticate with an API key created in Console → Settings → Developer. Keys are scoped; a scope does not imply any other scope. Full guide at https://circuvent.com/developers",
    "contact": { "name": "Circuvent Technologies", "url": "https://circuvent.com/contact" },
    "license": { "name": "Proprietary", "identifier": "LicenseRef-Circuvent" }
  },
  "servers": [{ "url": "https://api.circuvent.com", "description": "Production" }],
  "security": [{ "ApiKeyAuth": [] }],
...
  "paths": {
...
    "/v1/me": {
      "get": {
        "tags": ["Discovery"],
        "operationId": "getMe",
        "summary": "Identify this credential",
        "description": "Returns the account the key belongs to and the scopes it actually holds. The fastest way to diagnose a 403.",
        "security": [{ "ApiKeyAuth": ["devices:read"] }],
        "responses": {
          "200": {
            "description": "Credential details.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "account": {
                      "type": ["object", "null"],
                      "properties": {
                        "id": { "type": "integer" },
                        "email": { "type": "string" },
                        "name": { "type": "string" }
                      }
                    },
                    "deviceCount": { "type": "integer" },
                    "auth": {
                      "type": "object",
                      "properties": {
                        "method": { "type": "string", "enum": ["api_key", "session"] },
                        "keyName": { "type": "string" },
                        "env": { "type": "string", "enum": ["live", "test"] },
                        "scopes": { "type": "array", "items": { "type": "string" } }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/devices": {
      "get": {
        "tags": ["Devices"],
        "operationId": "listDevices",
        "summary": "List devices",
        "security": [{ "ApiKeyAuth": ["devices:read"] }],
        "parameters": [
          { "name": "room", "in": "query", "schema": { "type": "string" }, "description": "Exact room name." },
          { "name": "type", "in": "query", "schema": { "type": "string" }, "description": "Device type, e.g. home-hub." },
          { "name": "online", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": {
            "description": "The caller's devices.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "devices": { "type": "array", "items": { "$ref": "#/components/schemas/Device" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/v1/devices/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/DeviceId" }],
      "get": {
        "tags": ["Devices"],
        "operationId": "getDevice",
        "summary": "Get one device",
        "security": [{ "ApiKeyAuth": ["devices:read"] }],
        "responses": {
          "200": {
            "description": "The device.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "device": { "$ref": "#/components/schemas/Device" } }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["Devices"],
        "operationId": "updateDevice",
        "summary": "Rename, assign a room, or set favourite",
        "security": [{ "ApiKeyAuth": ["devices:write"] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "maxLength": 120 },
                  "room": { "type": "string", "maxLength": 80 },
                  "favorite": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated device.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "device": { "$ref": "#/components/schemas/Device" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/devices/{id}/commands": {
      "parameters": [{ "$ref": "#/components/parameters/DeviceId" }],
      "post": {
        "tags": ["Devices"],
        "operationId": "sendCommand",
        "summary": "Send a command",
        "description": "The body is forwarded to the device unchanged; the accepted fields are defined by that product's firmware. Responds 202 — the broker has accepted the command for delivery, not that the relay has closed. Observe the result by re-reading the device or taking a webhook.",
        "security": [{ "ApiKeyAuth": ["devices:control"] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "additionalProperties": true },
              "examples": {
                "hubChannel": { "summary": "Hub channel 1 on", "value": { "ch": 0, "on": true } },
                "allRelays": { "summary": "All four channels on", "value": { "relays": [true, true, true, true] } },
                "singleRelay": { "summary": "Plug or light", "value": { "power": true } },
                "dimmer": { "summary": "Dimmable light at 60%", "value": { "power": true, "brightness": 60 } },
                "fan": { "summary": "Fan speed 3", "value": { "speed": 3 } },
                "scene": { "summary": "Named hub scene", "value": { "scene": "movie" } }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted for delivery.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accepted": { "type": "boolean" },
                    "deviceId": { "type": "string" },
                    "command": { "type": "object", "additionalProperties": true }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/v1/devices/{id}/telemetry": {
      "parameters": [{ "$ref": "#/components/parameters/DeviceId" }],
      "get": {
        "tags": ["Telemetry"],
        "operationId": "getTelemetry",
        "summary": "Telemetry history",
        "security": [{ "ApiKeyAuth": ["telemetry:read"] }],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 100 } },
          { "name": "since", "in": "query", "schema": { "type": "string", "format": "date-time" } }
        ],
        "responses": {
          "200": {
            "description": "Samples, newest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deviceId": { "type": "string" },
                    "telemetry": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "at": { "type": "string", "format": "date-time" },
                          "data": { "type": "object", "additionalProperties": true }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/v1/rooms": {
      "get": {
        "tags": ["Organisation"],
        "operationId": "listRooms",
        "summary": "List rooms",
        "security": [{ "ApiKeyAuth": ["rooms:read"] }],
        "responses": {
          "200": {
            "description": "Rooms with device counts.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rooms": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "name": { "type": "string" },
                          "icon": { "type": "string" },
                          "sort": { "type": "integer" },
                          "deviceCount": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/scenes": {
      "get": {
        "tags": ["Organisation"],
        "operationId": "listScenes",
        "summary": "List scenes",
        "security": [{ "ApiKeyAuth": ["scenes:read"] }],
        "responses": {
          "200": {
            "description": "Scenes and their actions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scenes": { "type": "array", "items": { "$ref": "#/components/schemas/Scene" } }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "cvk_live_… / cvk_test_…",
        "description": "An API key from Console → Settings → Developer. `X-API-Key: <key>` is accepted as an alternative header. Keys are refused from a browser origin unless that origin is registered on the key."
      }
    },
    "parameters": {
      "DeviceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" },
        "description": "The device id, as shown in the console and used in its MQTT topics."
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Missing, invalid, expired or revoked key. `code` is one of key_invalid, key_expired, key_revoked, key_blocked.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "The key is valid but lacks the required scope (insufficient_scope) or was used from a disallowed browser origin (origin_not_allowed).",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    }
  }
}